+++ src/test/java/com/example/sijangtong/brianServiceTest/ServiceTest.java
@@ -170,7 +170,13 @@ public class ServiceTest {
   @Transactional
   public void deleteProduct() {
     Product product = productRepository.findById(198L).get();
-    List<OrderItem> orderItems = orderItemRepository.findByProduct(product);
+    Optional<OrderItem> oResult = orderItemRepository.findByProduct(product);
+
+    if (oResult.isPresent()) {
+      OrderItem orderItem = oResult.get();
+
+      orderItemRepository.delete(orderItem);
+    }
 
     productImgRepository.deleteByProduct(product);
 
@@ -230,10 +236,13 @@ public class ServiceTest {
       });
 
       // orderItem product 1:1로 바뀐거 확인하면 다시 바꾸기
-      List<OrderItem> orderItems = orderItemRepository.findByProduct(product);
-      orderItems.forEach(orderItem -> {
+      Optional<OrderItem> oResult = orderItemRepository.findByProduct(product);
+
+      if (oResult.isPresent()) {
+        OrderItem orderItem = oResult.get();
+
         orderItemRepository.delete(orderItem);
-      });
+      }
 
       productImgRepository.deleteByProduct(product);
